requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.109.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
openai==2.30.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.33.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.36.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
watchfiles==1.1.1
websockets==16.0
